# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Serviço Backend",
        "description": "API RESTful para gerenciamento de serviços comunitários, incluindo pessoas, perfis, missões, notícias e doações.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pessoas/{pessoaId}/enderecos": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Lista endereços de uma pessoa",
                "operationId": "cf6159e1b940772298f30d6514707a6e",
                "parameters": [
                    {
                        "name": "pessoaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de endereços",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/enderecos/{id}": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Busca endereço por ID",
                "operationId": "04f205c824b703870cfdbbbcf75cc3f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço encontrado"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Atualiza um endereço",
                "operationId": "5aac4f6f8753fbad5804363cd0f793f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cep": {
                                        "type": "string"
                                    },
                                    "logradouro": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "complemento": {
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "type": "string"
                                    },
                                    "principal": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Exclui um endereço",
                "operationId": "9f4f92c7f07c1b2f09355a725a9d51cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Endereço excluído"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            }
        },
        "/api/enderecos": {
            "post": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Cria um novo endereço",
                "operationId": "7503e124e0fd8144c6a9901d2e08bb19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pessoa_id",
                                    "cep",
                                    "logradouro",
                                    "numero",
                                    "bairro"
                                ],
                                "properties": {
                                    "pessoa_id": {
                                        "type": "integer"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "95900-000"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "example": "Rua Principal"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "example": "Apto 101"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "principal": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço criado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/enderecos/{id}/principal": {
            "patch": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Define endereço como principal",
                "operationId": "1c766ee52fcba373349c3faa724b1711",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço definido como principal"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            }
        },
...
    },
...
}